Added name mangling to async callback typedefs
In Franca, it is possible to overload return arguments of methods. In
this case, the callback for two methods with different return arguments
but equal names must not be the same. This now is solved via the
addition of name mangling to such method callbacks. Also eliminated
unused dependency (guava).

Change-Id: If3e262b815c139253a600d78a75dc816e96c4fc2